assistant2: Factor out `ContextStrip`
