+++ crates/assistant2/src/assistant.rs
@@ -3,6 +3,7 @@ mod assistant_panel;
 mod assistant_settings;
 mod context;
 mod context_picker;
+mod context_strip;
 mod inline_assistant;
 mod message_editor;
 mod prompts;
